Removed old depricated files
