--- src/Chirp.CLI.Client/Chirp.CLI.Client.csproj
@@ -1,26 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net9.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CsvHelper" Version="33.1.0" />
-    <PackageReference Include="docopt.net" Version="0.8.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.9">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.9" />
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.9" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="9.0.9" />
-    <ProjectReference Include="../Chirp.CSVDB/Chirp.CSVDB.csproj" />
-    <ProjectReference Include="..\Chirp.General\Chirp.General.csproj" />
-    <None Include="appsettings.Client.json" CopyToOutputDirectory="Always" />
-    <None Include="appsettings.Client.Test.json" CopyToOutputDirectory="Always" />
-  </ItemGroup>
-  
-</Project>
