--- src/Chirp.CLI.Client/UserInterface.cs
@@ -1,14 +1,0 @@
-using Chirp.General;
-
-namespace Chirp.CLI.Client;
-
-public static class UserInterface
-{
-    public static void PrintCheeps(IEnumerable<Cheep> cheeps)
-    {
-        foreach (var cheep in cheeps)
-        {
-            Console.WriteLine(cheep);
-        }
-    }
-}
