--- src/Chirp.CLI.Client/appsettings.Client.Test.json
@@ -1,6 +1,0 @@
-{
-  "AppSettings": {
-    "BaseURL": "http://localhost:",
-    "DefaultPort": "5000"
-  }
-}
