--- src/Chirp.CLI.Client/appsettings.Client.json
@@ -1,5 +1,0 @@
-{
-  "AppSettings": {
-    "BaseURL": "https://bdsagroup7chirpremotedb-h5efe2emb9hhcwad.swedencentral-01.azurewebsites.net//"
-  }
-}
