--- src/Chirp.CSVDBService/Chirp.CSVDBService.csproj
@@ -1,28 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net9.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <AssemblyName>Chirp.CSVDBService</AssemblyName>
-  </PropertyGroup>
-
-
-  <ItemGroup>
-    <ProjectReference Include="../Chirp.General/Chirp.General.csproj" />
-    <None Include="chirp_cli_db_Testing.csv" CopyToOutputDirectory="Always" />
-    <None Include="chirp_cli_db.csv" CopyToOutputDirectory="Always" />
-    <InternalsVisibleTo Include="../../tests/Chirp.Razor.Tests" />
-    <ProjectReference Include="..\Chirp.DBFacade\Chirp.DBFacade.csproj" />
-  </ItemGroup>
-
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.9">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="9.0.9" />
-  </ItemGroup>
-  
-</Project>
