--- src/Chirp.CSVDBService/appsettings.Services.Test.json
@@ -1,5 +1,0 @@
-{
-  "AppSettings": {
-    "DatabaseLocation": "chirp_cli_db_Testing.csv"
-  }
-}
