--- src/Chirp.CSVDBService/appsettings.Services.json
@@ -1,5 +1,0 @@
-{
-  "AppSettings": {
-    "DatabaseLocation": "chirp_cli_db.csv"
-  }
-}
