--- tests/Chirp.CSVDB.Tests/data/empty.csv
@@ -1,1 +1,0 @@
-﻿Author,Message,Timestamp
