--- tests/Chirp.CSVDB.Tests/data/noHeaderAlt.csv
@@ -1,3 +1,0 @@
-﻿ropf,"Hello students!",1690891760
-adho,"Welcome to the course!",1690978778
-adho,"I hope you had a good summer.",1690979858
