--- tests/Chirp.CSVDB.Tests/data/nullRecordInMiddle.csv
@@ -1,10 +1,0 @@
-﻿Author,Message,Timestamp
-ropf,"Hello, BDSA students!",1690891760
-adho,"Welcome to the course!",1690978778
-adho,"I hope you had a good summer.",1690979858
-ropf,"Cheeping cheeps on Chirp :)",1690981487
-,,
-louis,"Hello, you rascal!",1756464518
-bikzi,"I can't believe I am actually cheeping!",1757066819
-bikzi,"maybe I can believe I am actually cheeping...",1757067430
-pines,"wow, so this is what it feels like to cheep!",1757242423
