--- tests/Chirp.CSVDB.Tests/data/oneNullRecord.csv
@@ -1,2 +1,0 @@
-﻿Author,Message,Timestamp
-,,
