--- tests/Chirp.CSVDB.Tests/data/twoRecordsOnOneLine.csv
@@ -1,2 +1,0 @@
-﻿Author,Message,Timestamp
-ropf,"Hello BDSA students!",1690891760,adho,"Welcome to the course!",1690978778
