+++ tests/Chirp.Razor.Tests/Chirp.Razor.Test.csproj
@@ -18,7 +18,6 @@
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.12.0" />
         <PackageReference Include="xunit" Version="2.9.2" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2" />
-        <ProjectReference Include="..\..\src\Chirp.CSVDBService\Chirp.CSVDBService.csproj" />
         
         <ProjectReference Include="..\..\src\Chirp.General\Chirp.General.csproj" />
         <ProjectReference Include="..\..\src\Chirp.Razor\Chirp.Razor.csproj" />
